+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 2.16 branch
 ### 2.16.0
 * it uses `CollectionInterface::first()` instead of the global `firstValue()`;
+* added `StaticPage::_getSlug() method`;
 * renamed `/ContactForm.php` as `ContactUsForm` and `ContactFormMailer` as
     `ContactUsMailer`. This creates less confusion with the classes names;
 * simplified the name of some methods of the `BaseUpdateShell` class.